HOC of controls (#58)
* Completion up

* Labels module to HOC function

* Label Mixin
